Fix exposure and ontime fits keys values with tis cameras.

+++ src/CameraGigeTis.h
@@ -97,7 +97,7 @@
 
             int mFrameCounter;
             int mGain;
-            int mExposure;
+            double mExposure;
             double mFPS;
             CamBitDepth mImgDepth;
             int mSaturateVal;
