+++ src/Frame.cpp
@@ -34,7 +34,7 @@
 
 #include "Frame.h"
 
-Frame::Frame(Mat capImg, int g, int e, string acquisitionDate):
+Frame::Frame(Mat capImg, int g, double e, string acquisitionDate):
 mExposure(e), mGain(g), mFileName("noFileName"), mFrameRemaining(0),
 mFrameNumber(0), mFps(0), mBitDepth(MONO_8), mSaturatedValue(255) {
